requirements.txt
nbtlib
pyyaml
paramiko>=3.0.0